crm_account_management: move QC administrative-region (AC2) to new l10n_ca_qc_crm_account_management module
Per review feedback, the Quebec-specific administrative-region logic (qc_regions
helper, qc_administrative_region stored computed field, group-by/filter) is moved
out of the shared community module crm_account_management into a new Quebec
localization module l10n_ca_qc_crm_account_management. The province-agnostic
address fields (AC1) remain in crm_account_management. No data migration: this PR
is unmerged, so the AC2 code has never been released.
