+++ l10n_ca_qc_crm_account_management/__init__.py
@@ -1,0 +1,1 @@
+from . import models
