+++ l10n_ca_qc_crm_account_management/models/__init__.py
@@ -1,2 +1,2 @@
+from . import qc_regions
 from . import organizational_unit
-from . import res_partner
