+++ l10n_ca_qc_crm_account_management/models/qc_regions.py
