+++ l10n_ca_qc_crm_account_management/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_qc_administrative_region
